added runtime config updated to the theme compilation and configuration updates
