+++ src/Storefront/Theme/ThemeMergedConfigBuilder.php
@@ -1,0 +1,404 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Storefront\Theme;
+
+use Shopware\Core\Framework\Context;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
+use Shopware\Core\Framework\DataAbstractionLayer\Exception\InconsistentCriteriaIdsException;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\Log\Package;
+use Shopware\Storefront\Theme\Exception\InvalidThemeConfigException;
+use Shopware\Storefront\Theme\Exception\ThemeException;
+
+/**
+ * @internal
+ */
+#[Package('framework')]
+class ThemeMergedConfigBuilder
+{
+    /**
+     * @internal
+     *
+     * @param EntityRepository<ThemeCollection> $themeRepository
+     */
+    public function __construct(
+        private readonly StorefrontPluginRegistry $extensionRegistry,
+        private readonly EntityRepository $themeRepository,
+    ) {
+    }
+
+    /**
+     * @throws InvalidThemeConfigException
+     * @throws ThemeException
+     * @throws InconsistentCriteriaIdsException
+     *
+     * @return array<string, mixed>
+     */
+    public function getThemeConfiguration(string $themeId, bool $translate, Context $context): array
+    {
+        $criteria = (new Criteria())
+            ->setTitle('theme-service::load-config');
+
+        $themes = $this->themeRepository->search($criteria, $context)->getEntities();
+
+        $theme = $themes->get($themeId);
+        if (!$theme) {
+            throw ThemeException::couldNotFindThemeById($themeId);
+        }
+
+        $baseTheme = $themes->filter(fn (ThemeEntity $themeEntry) => $themeEntry->getTechnicalName() === StorefrontPluginRegistry::BASE_THEME_NAME)->first();
+        if ($baseTheme === null) {
+            throw ThemeException::couldNotFindThemeByName(StorefrontPluginRegistry::BASE_THEME_NAME);
+        }
+
+        $baseThemeConfig = $this->mergeStaticConfig($baseTheme);
+
+        $themeConfigFieldFactory = new ThemeConfigFieldFactory();
+        $configFields = [];
+        $labels = array_replace_recursive($baseTheme->getLabels() ?? [], $theme->getLabels() ?? []);
+        $helpTexts = array_replace_recursive($baseTheme->getHelpTexts() ?? [], $theme->getHelpTexts() ?? []);
+
+        if ($theme->getParentThemeId()) {
+            foreach ($this->getParentThemes($themes, $theme) as $parentTheme) {
+                $configuredParentTheme = $this->mergeStaticConfig($parentTheme);
+                $baseThemeConfig = array_replace_recursive($baseThemeConfig, $configuredParentTheme);
+                $labels = array_replace_recursive($labels, $parentTheme->getLabels() ?? []);
+                $helpTexts = array_replace_recursive($helpTexts, $parentTheme->getHelpTexts() ?? []);
+            }
+        }
+
+        $configuredTheme = $this->mergeStaticConfig($theme);
+        $themeConfig = array_replace_recursive($baseThemeConfig, $configuredTheme);
+
+        foreach ($themeConfig['fields'] ?? [] as $name => $item) {
+            $configFields[$name] = $themeConfigFieldFactory->create($name, $item);
+            if (
+                isset($item['value'], $configuredTheme['fields'])
+                && \is_array($item['value'])
+                && \array_key_exists($name, $configuredTheme['fields'])
+            ) {
+                $configFields[$name]->setValue($configuredTheme['fields'][$name]['value']);
+            }
+        }
+
+        $configFields = json_decode((string) json_encode($configFields, \JSON_THROW_ON_ERROR), true, 512, \JSON_THROW_ON_ERROR);
+
+        if ($translate && !empty($labels)) {
+            $configFields = $this->translateLabels($configFields, $labels);
+        }
+
+        if ($translate && !empty($helpTexts)) {
+            $configFields = $this->translateHelpTexts($configFields, $helpTexts);
+        }
+
+        $themeConfig['fields'] = $configFields;
+        $themeConfig['currentFields'] = [];
+        $themeConfig['baseThemeFields'] = [];
+
+        foreach ($themeConfig['fields'] as $field => $fieldItem) {
+            $isInherited = $this->fieldIsInherited($field, $configuredTheme);
+            $themeConfig['currentFields'][$field]['isInherited'] = $isInherited;
+
+            if ($isInherited) {
+                $themeConfig['currentFields'][$field]['value'] = null;
+            } elseif (\array_key_exists('value', $fieldItem)) {
+                $themeConfig['currentFields'][$field]['value'] = $fieldItem['value'];
+            }
+
+            $isInherited = $this->fieldIsInherited($field, $baseThemeConfig);
+            $themeConfig['baseThemeFields'][$field]['isInherited'] = $isInherited;
+
+            if ($isInherited) {
+                $themeConfig['baseThemeFields'][$field]['value'] = null;
+            } elseif (\array_key_exists('value', $fieldItem) && isset($baseThemeConfig['fields'][$field]['value'])) {
+                $themeConfig['baseThemeFields'][$field]['value'] = $baseThemeConfig['fields'][$field]['value'];
+            }
+        }
+
+        return $themeConfig;
+    }
+
+    /**
+     * @return array<string, mixed>
+     */
+    public function getThemeConfigurationStructuredFields(string $themeId, bool $translate, Context $context): array
+    {
+        $mergedConfig = $this->getThemeConfiguration($themeId, $translate, $context)['fields'];
+
+        $translations = [];
+        if ($translate) {
+            $translations = $this->getTranslations($themeId, $context);
+            $mergedConfig = $this->translateLabels($mergedConfig, $translations);
+        }
+
+        $outputStructure = [];
+
+        foreach ($mergedConfig as $fieldName => $fieldConfig) {
+            $tab = $this->getTab($fieldConfig);
+            $tabLabel = $this->getTabLabel($tab, $translations);
+            $block = $this->getBlock($fieldConfig);
+            $blockLabel = $this->getBlockLabel($block, $translations);
+            $section = $this->getSection($fieldConfig);
+            $sectionLabel = $this->getSectionLabel($section, $translations);
+
+            // set default tab
+            $outputStructure['tabs']['default']['label'] = '';
+
+            // set labels
+            $outputStructure['tabs'][$tab]['label'] = $tabLabel;
+            $outputStructure['tabs'][$tab]['blocks'][$block]['label'] = $blockLabel;
+            $outputStructure['tabs'][$tab]['blocks'][$block]['sections'][$section]['label'] = $sectionLabel;
+
+            // add fields to sections
+            $outputStructure['tabs'][$tab]['blocks'][$block]['sections'][$section]['fields'][$fieldName] = [
+                'label' => $fieldConfig['label'],
+                'helpText' => $fieldConfig['helpText'] ?? null,
+                'type' => $fieldConfig['type'] ?? null,
+                'custom' => $fieldConfig['custom'],
+                'fullWidth' => $fieldConfig['fullWidth'],
+            ];
+        }
+
+        return $outputStructure;
+    }
+
+    /**
+     * @param array<string, ThemeEntity> $parentThemes
+     *
+     * @return array<string, ThemeEntity>
+     */
+    private function getParentThemes(ThemeCollection $themes, ThemeEntity $mainTheme, array $parentThemes = []): array
+    {
+        foreach ($this->getConfigInheritance($mainTheme) as $parentThemeName) {
+            $parentTheme = $themes->filter(fn (ThemeEntity $themeEntry) => $themeEntry->getTechnicalName() === str_replace('@', '', (string) $parentThemeName))->first();
+
+            if ($parentTheme instanceof ThemeEntity && !\array_key_exists($parentTheme->getId(), $parentThemes)) {
+                $parentThemes[$parentTheme->getId()] = $parentTheme;
+
+                if ($parentTheme->getParentThemeId()) {
+                    $parentThemes = $this->getParentThemes($themes, $mainTheme, $parentThemes);
+                }
+            }
+        }
+
+        if ($mainTheme->getParentThemeId()) {
+            $parentTheme = $themes->filter(fn (ThemeEntity $themeEntry) => $themeEntry->getId() === $mainTheme->getParentThemeId())->first();
+
+            if ($parentTheme instanceof ThemeEntity && !\array_key_exists($parentTheme->getId(), $parentThemes)) {
+                $parentThemes[$parentTheme->getId()] = $parentTheme;
+                if ($parentTheme->getParentThemeId()) {
+                    $parentThemes = $this->getParentThemes($themes, $mainTheme, $parentThemes);
+                }
+            }
+        }
+
+        return $parentThemes;
+    }
+
+    /**
+     * @return array<int, string>
+     */
+    private function getConfigInheritance(ThemeEntity $mainTheme): array
+    {
+        if (\is_array($mainTheme->getBaseConfig())
+            && \array_key_exists('configInheritance', $mainTheme->getBaseConfig())
+            && \is_array($mainTheme->getBaseConfig()['configInheritance'])
+            && !empty($mainTheme->getBaseConfig()['configInheritance'])
+        ) {
+            return $mainTheme->getBaseConfig()['configInheritance'];
+        }
+
+        if ($mainTheme->getTechnicalName() !== StorefrontPluginRegistry::BASE_THEME_NAME) {
+            return [
+                '@' . StorefrontPluginRegistry::BASE_THEME_NAME,
+            ];
+        }
+
+        return [];
+    }
+
+    /**
+     * @return array<string, mixed>
+     */
+    private function mergeStaticConfig(ThemeEntity $theme): array
+    {
+        $configuredTheme = [];
+
+        $pluginConfig = null;
+        if ($theme->getTechnicalName()) {
+            $pluginConfig = $this->extensionRegistry->getConfigurations()->getByTechnicalName($theme->getTechnicalName());
+        }
+
+        if ($pluginConfig !== null) {
+            $configuredTheme = $pluginConfig->getThemeConfig();
+        }
+
+        if ($theme->getBaseConfig() !== null) {
+            $configuredTheme = array_replace_recursive($configuredTheme ?? [], $theme->getBaseConfig());
+        }
+
+        if ($theme->getConfigValues() !== null) {
+            foreach ($theme->getConfigValues() as $fieldName => $configValue) {
+                if (\array_key_exists('value', $configValue)) {
+                    $configuredTheme['fields'][$fieldName]['value'] = $configValue['value'];
+                }
+            }
+        }
+
+        return $configuredTheme ?: [];
+    }
+
+    /**
+     * @param array<string, mixed> $fieldConfig
+     */
+    private function getTab(array $fieldConfig): string
+    {
+        $tab = 'default';
+
+        if (isset($fieldConfig['tab'])) {
+            $tab = $fieldConfig['tab'];
+        }
+
+        return $tab;
+    }
+
+    /**
+     * @param array<string, mixed> $fieldConfig
+     */
+    private function getBlock(array $fieldConfig): string
+    {
+        $block = 'default';
+
+        if (isset($fieldConfig['block'])) {
+            $block = $fieldConfig['block'];
+        }
+
+        return $block;
+    }
+
+    /**
+     * @param array<string, mixed> $fieldConfig
+     */
+    private function getSection(array $fieldConfig): string
+    {
+        $section = 'default';
+
+        if (isset($fieldConfig['section'])) {
+            $section = $fieldConfig['section'];
+        }
+
+        return $section;
+    }
+
+    /**
+     * @param array<string, mixed> $translations
+     */
+    private function getTabLabel(string $tabName, array $translations): string
+    {
+        if ($tabName === 'default') {
+            return '';
+        }
+
+        return $translations['tabs.' . $tabName] ?? $tabName;
+    }
+
+    /**
+     * @param array<string, mixed> $translations
+     */
+    private function getBlockLabel(string $blockName, array $translations): string
+    {
+        if ($blockName === 'default') {
+            return '';
+        }
+
+        return $translations['blocks.' . $blockName] ?? $blockName;
+    }
+
+    /**
+     * @param array<string, mixed> $translations
+     */
+    private function getSectionLabel(string $sectionName, array $translations): string
+    {
+        if ($sectionName === 'default') {
+            return '';
+        }
+
+        return $translations['sections.' . $sectionName] ?? $sectionName;
+    }
+
+    /**
+     * @param array<string, mixed> $themeConfiguration
+     * @param array<string, mixed> $translations
+     *
+     * @return array<string, mixed>
+     */
+    private function translateLabels(array $themeConfiguration, array $translations): array
+    {
+        foreach ($themeConfiguration as $key => &$value) {
+            $value['label'] = $translations['fields.' . $key] ?? $key;
+        }
+
+        return $themeConfiguration;
+    }
+
+    /**
+     * @param array<string, mixed> $themeConfiguration
+     * @param array<string, mixed> $translations
+     *
+     * @return array<string, mixed>
+     */
+    private function translateHelpTexts(array $themeConfiguration, array $translations): array
+    {
+        foreach ($themeConfiguration as $key => &$value) {
+            $value['helpText'] = $translations['fields.' . $key] ?? null;
+        }
+
+        return $themeConfiguration;
+    }
+
+    /**
+     * @return array<string, mixed>
+     */
+    private function getTranslations(string $themeId, Context $context): array
+    {
+        $theme = $this->themeRepository->search(new Criteria([$themeId]), $context)->getEntities()->first();
+        if (!$theme) {
+            throw ThemeException::couldNotFindThemeById($themeId);
+        }
+
+        $translations = $theme->getLabels() ?: [];
+
+        if ($theme->getTechnicalName() !== StorefrontPluginRegistry::BASE_THEME_NAME) {
+            $criteria = (new Criteria())
+                ->setTitle('theme-service::load-translations');
+
+            $themes = $this->themeRepository->search($criteria, $context)->getEntities();
+            foreach ($this->getParentThemes($themes, $theme) as $parentTheme) {
+                $parentTranslations = $parentTheme->getLabels() ?: [];
+                $translations = array_replace_recursive($parentTranslations, $translations);
+            }
+        }
+
+        return $translations;
+    }
+
+    /**
+     * @param array<string, mixed> $configuration
+     */
+    private function fieldIsInherited(string $fieldName, array $configuration): bool
+    {
+        if (!isset($configuration['fields'])) {
+            return true;
+        }
+
+        if (!\is_array($configuration['fields'])) {
+            return true;
+        }
+
+        if (!\array_key_exists($fieldName, $configuration['fields'])) {
+            return true;
+        }
+
+        return false;
+    }
+}
